Rename peer_states to simply states
Reflecting its now decoupled nature.
This commit separated from previous to keep the previous diff readable.
